vlm_finetuning/requirements.txt
absl-py==2.1.0
accelerate==1.1.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
anyio==4.6.2.post1
attrdict==2.0.1
attrs==24.2.0
autocommand==2.2.2
av==13.1.0
backports.tarfile==1.2.0
binpacking==1.5.2
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cpm-kernels==1.0.11
crcmod==1.7
cryptography==43.0.3
cycler==0.12.1
dacite==1.8.1
datasets==2.21.0
deepspeed==0.15.3
dill==0.3.8
distro==1.9.0
docstring_parser==0.16
einops==0.8.0
fastapi==0.115.4
ffmpy==0.4.0
filelock==3.16.1
flash_attn==2.6.3
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.6.1
future==1.0.0
gradio==5.5.0
gradio_client==1.4.2
grpcio==1.67.1
h11==0.14.0
hjson==3.1.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.0
inflate64==1.0.0
inflect==7.3.1
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jieba==0.42.1
Jinja2==3.1.4
jiter==0.7.0
jmespath==0.10.0
joblib==1.4.2
kiwisolver==1.4.7
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
modelscope==1.18.1
more-itertools==10.3.0
mpmath==1.3.0
ms-swift[all]
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
multivolumefile==0.2.3
networkx==3.4.2
ninja==1.11.1.1
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.54.3
orjson==3.10.11
oss2==2.19.1
packaging==24.1
pandas==2.2.3
peft==0.12.0
pillow==11.0.0
platformdirs==4.2.2
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
py-cpuinfo==9.0.0
py7zr==0.22.0
pyarrow==18.0.0
pyav==13.1.0
pybcj==1.0.2
pycparser==2.22
pycryptodome==3.21.0
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
pyppmd==1.1.0
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
pyzstd==0.16.2
qwen-vl-utils==0.0.8
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rouge==1.0.1
ruff==0.7.2
safehttpx==0.1.1
safetensors==0.4.5
scipy==1.14.1
semantic-version==2.10.0
sentencepiece==0.2.0
setuptools==75.3.0
shellingham==1.5.4
shtab==1.7.1
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.41.2
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
texttable==1.7.0
tiktoken==0.8.0
tokenizers==0.20.3
tomli==2.0.1
tomlkit==0.12.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
transformers @ git+https://github.com/huggingface/transformers.git@7bbc62474391aff64f63fcc064c975752d1fa4de
transformers-stream-generator==0.0.5
triton==3.1.0
trl==0.11.4
typeguard==4.3.0
typer==0.12.5
typing_extensions==4.12.2
tyro==0.8.14
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
websockets==12.0
Werkzeug==3.1.2
wheel==0.45.0
xxhash==3.5.0
yarl==1.17.1
zipp==3.20.2
